Coming soon
Addresses some of #25 and closes #31.

Created an interim landing page for the almanac. I have emulated
some of the structure of the main HA site, for sake of consistency.
I have included the talisman/CSP stuff, and have manually copied
across some images and styles for the placeholder, but haven't
included the full bootstrap.min.css.