Combined routes for I18n
I've combined the routes for I18n using default parameters. I've also moved language to a hash map, and created a default for year.

Finally, I created some of the templates for 'ja' so that I could test that the routing was working across languages as expected, I hope that's alright.