Refactored the validating of year and language into a decorator
Refactored the i18n validation into a dedicated decorator that will
respect the selection of language and year. It will also try to
read the user language out of the 'Accept-Language' header as far
as possible. I've left a *lot* of debug logging in place, so that
we can verify that it's doing the right thing. The only real compromise
has been to accept the year as a parameter for the splash page, which
I feel is reasonable.