add routing for scaffolding templates (#49)
* add routing for scaffolding templates

* Combined routes for I18n (#53)

* Combined routes for I18n

I've combined the routes for I18n using default parameters. I've also moved language to a hash map, and created a default for year.

Finally, I created some of the templates for 'ja' so that I could test that the routing was working across languages as expected, I hope that's alright.

* Oops, forgot to remove the unused SUPPORTED_LANGS variable.

* Refactored the validating of year and language into a decorator (#58)

* Refactored the validating of year and language into a decorator

Refactored the i18n validation into a dedicated decorator that will
respect the selection of language and year. It will also try to
read the user language out of the 'Accept-Language' header as far
as possible. I've left a *lot* of debug logging in place, so that
we can verify that it's doing the right thing. The only real compromise
has been to accept the year as a parameter for the splash page, which
I feel is reasonable.

* Renamed validate decorator

Renamed the `validate_i18n` decorator to `validate`, and moved
the `abort` import back to flask.

* Added tests for parsing 'Accept-Language'

Wrote some tests to ensure that the parse_accept_language logic
support all of the required scenarios. A notable omission is that
I haven't tested the decorator itself, yet, as it requires mocking
the http request.

* Added todos for the custom error page and renamed the test to follow a better convention.

* Be more permissive with defaulting args

Changed the decorator to be more permissive of the wrapped function's
args. It will inspect the args that a function accepts before trying to
set them on kwargs, so that we don't bloat the route functions
unnecessarily.

* Added detail on how to run the tests to the README.md