First draft of Fonts chapter
This adds a first draft of text, as well as images, for the fonts
chapter.

There's quite a bit more text, including some written by Jason, that
hasn't been folded into this text. But I think it makes sense to get
this in and then propose refinements on top of it.

Closes #902