2020 Fonts chapter (#1643)
* First draft of Fonts chapter

This adds a first draft of text, as well as images, for the fonts
chapter.

There's quite a bit more text, including some written by Jason, that
hasn't been folded into this text. But I think it makes sense to get
this in and then propose refinements on top of it.

Closes #902

* Optimised images with calibre/image-actions

* Tweak images

This is mostly renaming as requested. This also uses the new "popular
typefaces" data as generated by rick.

* Optimised images with calibre/image-actions

* add raph as contributor

* More text

This commit adds sections on WOFF (mime types generally), a paragraph on
cache partitioning, local vs external hosting, and popular fonts.

* Link for cache partitioning

* Better figures

* Fix first chart

* Better descriptions

* Fix typo

* Adding Jason’s bio

* Fixing typo

* Adding bits to the variable fonts section

* Last few figure descriptions

* Response to review

I've done my best to improve the text in response to thoughtful review
comments by @svgeesus.

* Address review feedback

I've done my best to respond to feedback from @RoelN

* Proper apostrophes and quotation marks

* Adding bits about service speed comparisons & PFE

* Switching back to straight quotes

* More polishing

A few improvements to the text, trying to accommodate all feedback.

* Sharper characterization of JP font size

I didn't get feedback on this from the PR, but think it's an improvement.

* Update src/content/en/2020/fonts.md

Adopt Roel's suggestion on EOT

Co-authored-by: Roel Nieskens <RoelN@users.noreply.github.com>

* One minor edit in the last paragraph

* Improve stats (merge AbbyTsai suggestion)

Co-authored-by: AbbyTsai <44442451+AbbyTsai@users.noreply.github.com>

* Improve stats (merge AbbyTsai suggestion)

Co-authored-by: AbbyTsai <44442451+AbbyTsai@users.noreply.github.com>

* Improve stats (merge AbbyTsai suggestion)

Co-authored-by: AbbyTsai <44442451+AbbyTsai@users.noreply.github.com>

* Minor wording tweaks

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Rick Viscomi <rviscomi@gmail.com>
Co-authored-by: Barry <barry@tunetheweb.com>
Co-authored-by: Jason Pamental <jason@rwt.io>
Co-authored-by: Roel Nieskens <RoelN@users.noreply.github.com>
Co-authored-by: AbbyTsai <44442451+AbbyTsai@users.noreply.github.com>